+++ src/sensors.h
@@ -3,6 +3,8 @@
 
 #include "platform.h"
 #define SENSOR_FREQ_DIVIDER 4096
+#define MAX_ADC_PINS 16
+#define MAX_KNOCK_SAMPLES 16
 
 typedef enum {
   SENSOR_MAP, /* Manifold Absolute Pressure */
@@ -21,14 +23,13 @@ typedef enum {
   SENSOR_NONE,
   SENSOR_ADC,
   SENSOR_FREQ,
-  SENSOR_DIGITAL,
+  SENSOR_PULSEWIDTH,
   SENSOR_CONST,
 } sensor_source;
 
 typedef enum {
   METHOD_LINEAR,
   METHOD_LINEAR_WINDOWED,
-  METHOD_TABLE,
   METHOD_THERM,
 } sensor_method;
 
@@ -52,14 +53,15 @@ struct sensor_input {
     float min;
     float max;
   } range;
+  float raw_min;
+  float raw_max;
   struct table *table;
-  float fixed_value;
   struct thermistor_config therm;
 
   float lag;
   struct {
-    uint32_t min;
-    uint32_t max;
+    float min;
+    float max;
     float fault_value;
   } fault_config;
   struct window {
@@ -73,13 +75,11 @@ struct sensor_input {
     degrees_t collection_start_angle;
   } window;
 
-  uint32_t raw_value;
-  float processed_value;
-  struct {
-    timeval_t last_sample_time;
-    float last_sample_value;
-    float value;
-  } derivative;
+  timeval_t time;
+  float raw_value;
+  float value;
+  float derivative;
+
   sensor_fault fault;
 };
 
@@ -90,6 +90,7 @@ typedef enum {
 } trigger_edge;
 
 typedef enum {
+  NONE,
   FREQ,    /* Only generate frequency information */
   TRIGGER, /* Use as source of decoder position */
 } freq_type;
@@ -118,10 +119,35 @@ struct knock_input {
   float value;
 };
 
-void sensors_process(sensor_source source);
+struct freq_update {
+  timeval_t time;
+  bool valid;
+
+  uint32_t pin;
+  float frequency; /* Hz */
+  float pulsewidth; /* seconds */
+};
+
+struct adc_update {
+  timeval_t time;
+  bool valid;
+  float values[MAX_ADC_PINS]; /* Pin values in V */
+};
+
+struct knock_update {
+  timeval_t time; /* Time of first sample in set */
+  bool valid;
+  int pin;
+  int n_samples;
+  float samples[MAX_KNOCK_SAMPLES];
+};
+
+void sensor_update_freq(const struct freq_update *);
+void sensor_update_adc(const struct adc_update *);
+void sensor_update_knock(const struct knock_update *);
+
 uint32_t sensor_fault_status();
 
-void knock_add_sample(struct knock_input *, float sample);
 void knock_configure(struct knock_input *);
 
 #ifdef UNITTEST